requirements.txt
streamlit
opencv-python-headless<3.4.17
torch
torchvision
ultralytics
numpy
Pillow 